Added separate preprocessing of data, save it in folder to then be used by model
